* Moving relevant code from BlockChainController to WalletController * Remove unnecessary IDispatcher from BlockChainController * Remove unnecessary IDialogHelper from WalletControler
All of this was to remove Circular Reference between BlockChainController and WalletController
