+++ neo-gui/UI/Messages/BlockchainPersistCompletMessage.cs
@@ -1,0 +1,6 @@
+﻿namespace Neo.UI.Messages
+{
+    public class BlockchainPersistCompletMessage
+    {
+    }
+}
