+++ neo-gui/UI/Messages/UpdateWalletMessage.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace Neo.UI.Messages
+{
+    public class UpdateWalletMessage
+    {
+        public TimeSpan PersistenceSpan { get; private set; }
+
+        public UpdateWalletMessage(TimeSpan persistenceSpan)
+        {
+            this.PersistenceSpan = persistenceSpan;
+        }
+    }
+}
